==> src/3_data_visualization.ipynb <==
import numpy as np
import pandas as pd

pd.options.display.max_columns = None

from matplotlib import pyplot as plt
import seaborn as sns


df_2022 = pd.read_csv("../data/2022-23_reg_season_clean.csv")
df_2023 = pd.read_csv("../data/2023-24_reg_season_clean.csv")


# main df
df = df_2023.copy()

print(df_2022.shape)
print(df_2023.shape)
df.head()

df['season_start_year'].unique()

df.columns

keeps = [
    "team",
    "opponent",
    "day_of_week",
    "date",
    "Start (ET)",
    "season_start_year",
    "is_playoff",
    "is_home_game",
    "game_time",
    "city_team_key",
    "abbreviation",
    "nickname",
    "timezone",
    "time_diff_from_et",
    "elevation",
    "Latitude",
    "Longitude",
    "city_team_key_opponent",
    "abbreviation_opponent",
    "nickname_opponent",
    "timezone_opponent",
    "time_diff_from_et_opponent",
    "elevation_opponent",
    "Latitude_opponent",
    "Longitude_opponent",
    "is_covid_bubble",
    "game_timediff_from_et",
    "game_lat",
    "game_lon",
    "prev_game_lon",
    "prev_game_lat",
    "home_team_abbr",
    "away_team_abbr",
    "game_id",
    "time_since_prev_game",
    "is_last_game_home",
    "is_2_games_ago_home",
    "is_3_games_ago_home",
    "is_4_games_ago_home",
    "is_5_games_ago_home",
    "is_6_games_ago_home",
    "is_7_games_ago_home",
    "num_away_last_5_games",
    "start_of_streak",
    "streak_id",
    "streak_counter",
    "incoming_away_game_streak",
    "days_since_1_games_ago",
    "days_since_2_games_ago",
    "days_since_3_games_ago",
    "days_since_4_games_ago",
    "days_since_5_games_ago",
    "days_since_6_games_ago",
    "days_since_7_games_ago",
    "num_games_last_4_days",
    "num_games_last_7_days",
    "distance_from_previous_game",
    "cumulative_distance",
    "is_back_to_back",
    "team_opponent",
    "abbreviation_opponent.1",
    "time_since_prev_game_opponent",
    "num_games_last_7_days_opponent",
    "distance_from_previous_game_opponent",
    "cumulative_distance_opponent",
    "hours_of_rest_difference",
    "games_last_7_diff",
    "prev_game_distance_diff",
    "cumulative_distance_diff",
]

# potential new features to add

# number of 4 in 5 sets by season

# number of 5 in 7 sets by season

# number of back-to-backs by season

plt.figure(figsize=(8,5))
# temp_df = df
temp_df = df[
    (df['team_abbreviation'] == "GSW")
    # & (df['season_start_year'] == 2018)
#     (df['cumulative_distance'] > 3000)
#     & (df['incoming_away_game_streak'] > 2)
]
sns.regplot(x = "cumulative_distance", y="point_difference", data=temp_df, line_kws={"color":"red"})
plt.grid()


plt.figure(figsize=(8,5))


sns.countplot(x = "incoming_away_game_streak", hue="WL", data=temp_df)
plt.grid()


temp_df = df[df['team_abbreviation'] == "ATL"]
print(temp_df.shape)
temp_df.head()
