==> src/11_hardest_pre-cup_schedule.ipynb <==
import numpy as np
import pandas as pd

pd.options.display.max_columns = None
pd.options.display.max_rows = None

from matplotlib import pyplot as plt
import seaborn as sns

CR_COLOR = "#0079FF"


df_2022 = pd.read_csv("../data/2022-23_reg_season_clean.csv")
df_2023 = pd.read_csv("../data/2023-24_reg_season_clean.csv")


# main df
# df = df_2022.copy()
df = df_2023.copy()

# convert to date type
df["game_time"] = pd.to_datetime(df["game_time"])
df["date"] = pd.to_datetime(df["date"])

print(df_2022.shape)
print(df_2023.shape)
df.head()

# mark the week of season
df['week_number'] = df['date'].dt.isocalendar().week
df['week_of_season'] = np.where(df['week_number'] >= 43, df['week_number'] - 42, df['week_number'] + 1 +(df['week_number'].max() - df['week_number'][0]))

df['team_game_id'] = df.groupby('abbreviation').cumcount() + 1
df.head()

# mark which games are before and after the TBD stage of the nba cup
after_cup_dict = df[df['date'].gt("2023-12-02")].groupby('abbreviation').first()['team_game_id'].to_dict()

df['after_id'] = df['abbreviation'].map(after_cup_dict)
df['is_first_after_cup'] = np.where(df['team_game_id'] == df['after_id'], 1, 0)
# df['distance_from_previous_game'] = np.where(df['is_first_after_cup'] == 1, 0, df['distance_from_previous_game'])
# df[['team', 'date', 'team_game_id',"is_home_game", "opponent", 'distance_from_previous_game', 'is_first_after_cup']].head(30)

# after_cup_dict

df.head(3)

df['is_post_nba_cup'].unique()

# df.columns

assert False

# total travel for each team (simple)
piv = df.pivot_table(
    index=["season_start_year","abbreviation", "team"],
    values="distance_from_previous_game",
    aggfunc="sum"
    
).sort_values(by="distance_from_previous_game").reset_index()
piv['team'] = piv['team'].str.split(" ").str[-1]
print(piv.shape)
# piv
# ax = piv.plot(
#     kind="bar",
#     x="abbreviation",
#     legend=False,
#     title="Total Distance traveled by each team for the 2023 Season",
#     ylabel="Miles",
#     xlabel="Team",
#     # grid='y',
#     figsize=(8,4)
# )
# ax.grid(axis="y")
piv.head(2)
